Make sure /data partition exists.

+++ openstack-hao/install-compute-node.sh
@@ -137,7 +137,8 @@ EOF
     service nova-compute restart; sleep 1
 
     # Move local nova storage from / to /data partition
-    if [ ! -e /data/nova ] && [ ! -L /var/lib/nova ]; then
+    if cut -d" " -f2 /proc/mounts | egrep -qs '^/data$' && \
+        [ ! -e /data/nova ] && [ ! -L /var/lib/nova ]; then
         cp -a /var/lib/nova /data
         rm -rf /var/lib/nova
         ln -s /data/nova /var/lib
